requirements.txt
streamlit>=1.41.0
anthropic>=0.43.0
PyMuPDF>=1.25.3
Pillow>=11.1.0
python-dotenv>=1.0.1
extra-streamlit-components>=0.1.71
